+++ hahomematic/platforms/custom/climate.py
@@ -10,7 +10,7 @@ from collections.abc import Mapping
 from datetime import datetime, timedelta
 from enum import IntEnum, StrEnum
 import logging
-from typing import Any, Final
+from typing import Any, Final, cast
 
 from hahomematic.const import HmPlatform, ParamsetKey
 from hahomematic.exceptions import ClientException, HaHomematicException, ValidationException
@@ -41,7 +41,7 @@ _PARTY_DATE_FORMAT: Final = "%Y_%m_%d %H:%M"
 _PARTY_INIT_DATE: Final = "2000_01_01 00:00"
 _RAW_SCHEDULE_DICT = dict[str, float | int]
 _TEMP_CELSIUS: Final = "°C"
-SCHEDULE_ENTRY_RANGE: Final = range(1, 13)
+SCHEDULE_SLOT_RANGE: Final = range(1, 14)
 SCHEDULE_TIME_RANGE: Final = range(1441)
 HM_PRESET_MODE_PREFIX: Final = "week_program_"
 
@@ -105,11 +105,15 @@ class HmPresetMode(StrEnum):
     WEEK_PROGRAM_6 = "week_program_6"
 
 
-class ScheduleEntryType(StrEnum):
+class ScheduleSlotType(StrEnum):
     """Enum for climate item type."""
 
     ENDTIME = "ENDTIME"
+    STARTTIME = "STARTTIME"
     TEMPERATURE = "TEMPERATURE"
+
+
+RELEVANT_SLOT_TYPES: Final = (ScheduleSlotType.ENDTIME, ScheduleSlotType.TEMPERATURE)
 
 
 class ScheduleProfile(StrEnum):
@@ -135,7 +139,8 @@ class ScheduleWeekday(StrEnum):
     SUNDAY = "SUNDAY"
 
 
-WEEKDAY_DICT = dict[int, dict[ScheduleEntryType, int | float]]
+SIMPLE_WEEKDAY_LIST = list[dict[ScheduleSlotType, str | float]]
+WEEKDAY_DICT = dict[int, dict[ScheduleSlotType, str | float]]
 PROFILE_DICT = dict[ScheduleWeekday, WEEKDAY_DICT]
 _SCHEDULE_DICT = dict[ScheduleProfile, PROFILE_DICT]
 
@@ -334,7 +339,7 @@ class BaseClimateEntity(CustomEntity):
                 f"Schedule is not supported by device {self._device.name}"
             ) from cex
 
-        for line, entry_value in raw_schedule.items():
+        for line, slot_value in raw_schedule.items():
             if not line.startswith("P"):
                 continue
             line_split = line.split("_")
@@ -344,19 +349,19 @@ class BaseClimateEntity(CustomEntity):
             _profile = ScheduleProfile(p)
             if profile and profile != _profile:
                 continue
-            _entry_type = ScheduleEntryType(et)
+            _slot_type = ScheduleSlotType(et)
             _weekday = ScheduleWeekday(w)
             if weekday and weekday != _weekday:
                 continue
-            _entry_no = int(no)
+            _slot_no = int(no)
 
             _add_to_schedule_data(
                 schedule_data=schedule_data,
                 profile=_profile,
                 weekday=_weekday,
-                entry_no=_entry_no,
-                entry_type=_entry_type,
-                entry_value=entry_value,
+                slot_no=_slot_no,
+                slot_type=_slot_type,
+                slot_value=slot_value,
             )
 
         return schedule_data
@@ -367,15 +372,15 @@ class BaseClimateEntity(CustomEntity):
         self._validate_profile(profile=profile, profile_data=profile_data)
         schedule_data: _SCHEDULE_DICT = {}
         for weekday, weekday_data in profile_data.items():
-            for entry_no, entry in weekday_data.items():
-                for entry_type, entry_value in entry.items():
+            for slot_no, slot in weekday_data.items():
+                for slot_type, slot_value in slot.items():
                     _add_to_schedule_data(
                         schedule_data=schedule_data,
                         profile=profile,
                         weekday=weekday,
-                        entry_no=entry_no,
-                        entry_type=entry_type,
-                        entry_value=entry_value,
+                        slot_no=slot_no,
+                        slot_type=slot_type,
+                        slot_value=slot_value,
                     )
         await self._client.put_paramset(
             channel_address=self._channel.address,
@@ -387,24 +392,93 @@ class BaseClimateEntity(CustomEntity):
     async def set_profile_weekday(
         self, profile: ScheduleProfile, weekday: ScheduleWeekday, weekday_data: WEEKDAY_DICT
     ) -> None:
-        """Set a profile to device."""
+        """Store a profile to device."""
         self._validate_profile_weekday(profile=profile, weekday=weekday, weekday_data=weekday_data)
         schedule_data: _SCHEDULE_DICT = {}
-        for entry_no, entry in weekday_data.items():
-            for entry_type, entry_value in entry.items():
+        for slot_no, slot in weekday_data.items():
+            for slot_type, slot_value in slot.items():
                 _add_to_schedule_data(
                     schedule_data=schedule_data,
                     profile=profile,
                     weekday=weekday,
-                    entry_no=entry_no,
-                    entry_type=entry_type,
-                    entry_value=entry_value,
+                    slot_no=slot_no,
+                    slot_type=slot_type,
+                    slot_value=slot_value,
                 )
         await self._client.put_paramset(
             channel_address=self._channel.address,
             paramset_key=ParamsetKey.MASTER,
             values=_get_raw_paramset(schedule_data=schedule_data),
         )
+
+    @service()
+    async def set_profile_weekday_simple(
+        self,
+        profile: ScheduleProfile,
+        weekday: ScheduleWeekday,
+        base_temperature: float,
+        simple_weekday_list: SIMPLE_WEEKDAY_LIST,
+    ) -> None:
+        """Store a simple profile to device."""
+        weekday_data = self._validate_and_convert_simple_to_weekday(
+            base_temperature=base_temperature, simple_weekday_list=simple_weekday_list
+        )
+        await self.set_profile_weekday(profile=profile, weekday=weekday, weekday_data=weekday_data)
+
+    def _validate_and_convert_simple_to_weekday(
+        self, base_temperature: float, simple_weekday_list: SIMPLE_WEEKDAY_LIST
+    ) -> WEEKDAY_DICT:
+        """Convert weekday dict to simple weekday list."""
+        if not self.min_temp <= base_temperature <= self.max_temp:
+            raise ValidationException(
+                f"VALIDATE_PROFILE_SIMPLE: Base temperature {base_temperature} not in valid range (min: {self.min_temp}, "
+                f"max: {self.max_temp})"
+            )
+
+        weekday_data: WEEKDAY_DICT = {}
+        sorted_simple_weekday_list = _sort_simple_weekday_list(
+            simple_weekday_list=simple_weekday_list
+        )
+        previous_endtime = "00:00"
+        slot_no = 1
+        for slot in sorted_simple_weekday_list:
+            if (starttime := slot.get(ScheduleSlotType.STARTTIME)) is None:
+                raise ValidationException("VALIDATE_PROFILE_SIMPLE: STARTTIME is missing.")
+            if (endtime := slot.get(ScheduleSlotType.ENDTIME)) is None:
+                raise ValidationException("VALIDATE_PROFILE_SIMPLE: ENDTIME is missing.")
+            if (temperature := slot.get(ScheduleSlotType.TEMPERATURE)) is None:
+                raise ValidationException("VALIDATE_PROFILE_SIMPLE: TEMPERATURE is missing.")
+
+            if _convert_time_str_to_minutes(str(starttime)) < _convert_time_str_to_minutes(
+                previous_endtime
+            ):
+                raise ValidationException(
+                    f"VALIDATE_PROFILE: Timespans are overlapping with a previous slot for starttime: {starttime} / endtime: {endtime}"
+                )
+
+            if not self.min_temp <= float(temperature) <= self.max_temp:
+                raise ValidationException(
+                    f"VALIDATE_PROFILE: Temperature {temperature} not in valid range (min: {self.min_temp}, "
+                    f"max: {self.max_temp}) for starttime: {starttime} / endtime: {endtime}"
+                )
+
+            if _convert_time_str_to_minutes(str(starttime)) > _convert_time_str_to_minutes(
+                previous_endtime
+            ):
+                weekday_data[slot_no] = {
+                    ScheduleSlotType.ENDTIME: starttime,
+                    ScheduleSlotType.TEMPERATURE: base_temperature,
+                }
+                slot_no += 1
+
+            weekday_data[slot_no] = {
+                ScheduleSlotType.ENDTIME: endtime,
+                ScheduleSlotType.TEMPERATURE: temperature,
+            }
+            previous_endtime = str(endtime)
+            slot_no += 1
+
+        return _fillup_weekday_data(base_temperature=base_temperature, weekday_data=weekday_data)
 
     def _validate_profile(self, profile: ScheduleProfile, profile_data: PROFILE_DICT) -> None:
         """Validate the profile."""
@@ -425,34 +499,36 @@ class BaseClimateEntity(CustomEntity):
     ) -> None:
         """Validate the profile weekday."""
         previous_endtime = 0
-        for no in SCHEDULE_ENTRY_RANGE:
+        for no in SCHEDULE_SLOT_RANGE:
             if no not in weekday_data:
                 raise ValidationException(
-                    f"VALIDATE_PROFILE: Entry no {no} is missing in profile: {profile}/weekday: {weekday}"
+                    f"VALIDATE_PROFILE: slot no {no} is missing in profile: {profile} / weekday: {weekday}"
                 )
-            entry = weekday_data[no]
-            for entry_type in ScheduleEntryType:
-                if entry_type not in entry:
+            slot = weekday_data[no]
+            for slot_type in RELEVANT_SLOT_TYPES:
+                if slot_type not in slot:
                     raise ValidationException(
-                        f"VALIDATE_PROFILE: Entry type {entry_type} is missing in profile: "
-                        f"{profile}/weekday: {weekday}/entry_no: {no}"
+                        f"VALIDATE_PROFILE: slot type {slot_type} is missing in profile: "
+                        f"{profile} / weekday: {weekday} / slot_no: {no}"
                     )
-                temperature = weekday_data[no][ScheduleEntryType.TEMPERATURE]
+                temperature = float(weekday_data[no][ScheduleSlotType.TEMPERATURE])
                 if not self.min_temp <= temperature <= self.max_temp:
                     raise ValidationException(
                         f"VALIDATE_PROFILE: Temperature {temperature} not in valid range (min: {self.min_temp}, "
-                        f"max: {self.max_temp}) for profile: {profile}/weekday: {weekday}/entry_no: {no}"
+                        f"max: {self.max_temp}) for profile: {profile} / weekday: {weekday} / slot_no: {no}"
                     )
-                if endtime := int(weekday_data[no][ScheduleEntryType.ENDTIME]):
+
+                endtime_str = str(weekday_data[no][ScheduleSlotType.ENDTIME])
+                if endtime := _convert_time_str_to_minutes(time_str=endtime_str):
                     if endtime not in SCHEDULE_TIME_RANGE:
                         raise ValidationException(
-                            f"VALIDATE_PROFILE: Time {endtime} must be between {SCHEDULE_TIME_RANGE.start} and "
-                            f"{SCHEDULE_TIME_RANGE.stop - 1} for profile: {profile}/weekday: {weekday}/entry_no: {no}"
+                            f"VALIDATE_PROFILE: Time {endtime_str} must be between {_convert_minutes_to_time_str(minutes=SCHEDULE_TIME_RANGE.start)} and "
+                            f"{_convert_minutes_to_time_str(minutes=SCHEDULE_TIME_RANGE.stop - 1)} for profile: {profile} / weekday: {weekday} / slot_no: {no}"
                         )
                     if endtime < previous_endtime:
                         raise ValidationException(
-                            f"VALIDATE_PROFILE: Time sequence must be rising. {endtime} is lower than the previous "
-                            f"value {previous_endtime} for profile: {profile}/weekday: {weekday}/entry_no: {no}"
+                            f"VALIDATE_PROFILE: Time sequence must be rising. {endtime_str} is lower than the previous "
+                            f"value {_convert_minutes_to_time_str(minutes=previous_endtime)} for profile: {profile} / weekday: {weekday} / slot_no: {no}"
                         )
                 previous_endtime = endtime
 
@@ -813,15 +889,60 @@ class CeIpThermostat(BaseClimateEntity):
         return profiles
 
 
+def _convert_minutes_to_time_str(minutes: int) -> str:
+    """Convert minutes to a time string."""
+    try:
+        return f"{minutes//60:0=2}:{minutes%60:0=2}"
+    except Exception as ex:
+        raise ValidationException(ex) from ex
+
+
+def _convert_time_str_to_minutes(time_str: str) -> int:
+    """Convert minutes to a time string."""
+    try:
+        h, m = time_str.split(":")
+        return (int(h) * 60) + int(m)
+    except Exception as ex:
+        raise ValidationException(
+            f"Failed to convert time {time_str}. Format must be hh:mm."
+        ) from ex
+
+
+def _sort_simple_weekday_list(simple_weekday_list: SIMPLE_WEEKDAY_LIST) -> SIMPLE_WEEKDAY_LIST:
+    """Sort simple weekday list."""
+    simple_weekday_dict = sorted(
+        {
+            _convert_time_str_to_minutes(str(slot[ScheduleSlotType.STARTTIME])): slot
+            for slot in simple_weekday_list
+        }.items()
+    )
+    return [slot[1] for slot in simple_weekday_dict]
+
+
+def _fillup_weekday_data(base_temperature: float, weekday_data: WEEKDAY_DICT) -> WEEKDAY_DICT:
+    """Fillup weekday data."""
+    for slot_no in SCHEDULE_SLOT_RANGE:
+        if slot_no not in weekday_data:
+            weekday_data[slot_no] = {
+                ScheduleSlotType.ENDTIME: "24:00",
+                ScheduleSlotType.TEMPERATURE: base_temperature,
+            }
+
+    return weekday_data
+
+
 def _get_raw_paramset(schedule_data: _SCHEDULE_DICT) -> _RAW_SCHEDULE_DICT:
     """Return the raw paramset."""
     raw_paramset: _RAW_SCHEDULE_DICT = {}
     for profile, profile_data in schedule_data.items():
         for weekday, weekday_data in profile_data.items():
-            for entry_no, entry in weekday_data.items():
-                for entry_type, entry_value in entry.items():
-                    raw_paramset[f"{str(profile)}_{str(entry_type)}_{str(weekday)}_{entry_no}"] = (
-                        entry_value
+            for slot_no, slot in weekday_data.items():
+                for slot_type, slot_value in slot.items():
+                    raw_value: float | int = cast(float | int, slot_value)
+                    if slot_type == ScheduleSlotType.ENDTIME and isinstance(slot_value, str):
+                        raw_value = _convert_time_str_to_minutes(slot_value)
+                    raw_paramset[f"{str(profile)}_{str(slot_type)}_{str(weekday)}_{slot_no}"] = (
+                        raw_value
                     )
     return raw_paramset
 
@@ -830,19 +951,21 @@ def _add_to_schedule_data(
     schedule_data: _SCHEDULE_DICT,
     profile: ScheduleProfile,
     weekday: ScheduleWeekday,
-    entry_no: int,
-    entry_type: ScheduleEntryType,
-    entry_value: float | int,
+    slot_no: int,
+    slot_type: ScheduleSlotType,
+    slot_value: str | float,
 ) -> None:
-    """Add or update schedule entry."""
+    """Add or update schedule slot."""
     if profile not in schedule_data:
         schedule_data[profile] = {}
     if weekday not in schedule_data[profile]:
         schedule_data[profile][weekday] = {}
-    if entry_no not in schedule_data[profile][weekday]:
-        schedule_data[profile][weekday][entry_no] = {}
-    if entry_type not in schedule_data[profile][weekday][entry_no]:
-        schedule_data[profile][weekday][entry_no][entry_type] = entry_value
+    if slot_no not in schedule_data[profile][weekday]:
+        schedule_data[profile][weekday][slot_no] = {}
+    if slot_type not in schedule_data[profile][weekday][slot_no]:
+        if slot_type == ScheduleSlotType.ENDTIME and isinstance(slot_value, int):
+            slot_value = _convert_minutes_to_time_str(slot_value)
+        schedule_data[profile][weekday][slot_no][slot_type] = slot_value
 
 
 def make_simple_thermostat(